# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AKSIS Core API
  version: 1.13.0
  description: >-
    Integrated Phase 03-15 REST contract covering authentication and tenant
    context, schools/classes/students, academic periods and student history,
    card lifecycle, and secure device provisioning/runtime operations.
  x-aksis-phases:
    - phase: 3
      resources: [Auth, Schools, Classes, Students]
    - phase: 4
      resources: [Academic Years, Student History, Cards]
    - phase: 5
      resources: [Devices]
    - phase: 6
      resources: [Gate Attendance, Offline Sync]
    - phase: 7
      resources: [Admin Context, Attendance Read Model, Device Inventory]
    - phase: 8
      resources: [Parent Linking, Children, Parent Timeline]
    - phase: 9
      resources: [Waste Transactions, Ranking, Summary]
    - phase: 10
      resources: [Extracurriculars, Sessions, Members, Attendance]
    - phase: 11
      resources: [Library Visits, Offline Sync, Class Summary]
    - phase: 12
      resources: [LED Content, Gateway State, Heartbeat, Acknowledgement]
    - phase: 13
      resources: [Command Center, Today Aggregate]
    - phase: 14
      resources: [Card Writer Jobs, Station Lease, Read-back Verification]
    - phase: 15
      resources: [Reporting, Audit, Monitoring]
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
    schoolId: []
paths:
...
  /classes:
    get:
      tags: [Classes]
      operationId: listClasses
      summary: List classes in the selected tenant
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: Paginated classes
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    required: [data, meta]
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Class' }
                      meta: { $ref: '#/components/schemas/PaginationMeta' }
        '401': { $ref: '#/components/responses/AuthRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Classes]
      operationId: createClass
      summary: Create a class
      description: Requires `academic.manage`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClassCreate' }
      responses:
        '201':
          description: Created class
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClassResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /cards:
    get:
      tags: [Cards]
      operationId: listCards
      summary: List cards for the active tenant
      description: Requires `card.read`.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - { name: student_id, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/CardStatus' } }
      responses:
        '200': { description: Paginated cards, content: { application/json: { schema: { $ref: '#/components/schemas/CardListResponse' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Cards]
      operationId: createCard
      deprecated: true
      summary: Rejected; create and release a printed production batch instead
      responses:
        '409': { $ref: '#/components/responses/Conflict' }
  /cards/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    get:
      tags: [Cards]
      operationId: getCard
      summary: Get a card
      responses:
        '200': { description: Card details, content: { application/json: { schema: { $ref: '#/components/schemas/CardResponse' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Cards]
      operationId: updateCardStatus
      summary: Change card status or expiration
      description: Requires `card.manage`.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CardUpdate' } } } }
      responses:
        '200': { description: Updated card, content: { application/json: { schema: { $ref: '#/components/schemas/CardResponse' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Cards]
      operationId: blockCard
      deprecated: true
      summary: Rejected; use PATCH with status and reason
      responses:
        '409': { $ref: '#/components/responses/Conflict' }
...
  /platform/production/candidates:
    get:
      tags: [Platform]
      operationId: cardProductionCandidates
      summary: "Paginated active students; school_id required; class_id and search optional"
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/production/classes:
    get:
      tags: [Platform]
      operationId: cardProductionClasses
      summary: "Active classes with academic years; school_id required"
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/production/batches:
    get:
      tags: [Platform]
      operationId: cardProductionBatches
      summary: "Paginated print batches"
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
    post:
      tags: [Platform]
      operationId: createCardProductionBatch
      summary: "Create inactive card identities and immutable print snapshots"
      security: [{ bearerAuth: [] }]
      requestBody: {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["id", "school_id", "student_ids"], "properties": {"id": {"type": "string", "format": "uuid"}, "school_id": {"type": "string", "format": "uuid"}, "student_ids": {"type": "array", "minItems": 1, "maxItems": 200, "items": {"type": "string", "format": "uuid"}}, "class_id": {"type": ["string", "null"], "format": "uuid"}, "replaces_card_id": {"type": ["string", "null"], "format": "uuid"}}}}}}
      responses: { '201': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/production/batches/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    get:
      tags: [Platform]
      operationId: cardProductionBatchCards
      summary: "Cards, opaque QR and snapshots for a batch"
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/production/batches/{id}/action:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    post:
      tags: [Platform]
      operationId: cardProductionBatchAction
      summary: "Record physical print, reprint or release after QC"
      security: [{ bearerAuth: [] }]
      requestBody: {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["event_id", "action", "reason"], "properties": {"event_id": {"type": "string", "format": "uuid"}, "action": {"type": "string", "enum": ["PRINTED", "REPRINTED", "RELEASED"]}, "reason": {"type": "string", "minLength": 3, "maxLength": 500}}}}}}
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/production/jobs:
    get:
      tags: [Platform]
      operationId: cardProductionJobs
      summary: "Reconcile expired leases and return paginated jobs and attempt logs"
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/production/jobs/{id}/action:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    post:
      tags: [Platform]
      operationId: cardProductionJobAction
      summary: "Explicit retry or cancellation with audit reason"
      security: [{ bearerAuth: [] }]
      requestBody: {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["action", "reason"], "properties": {"action": {"type": "string", "enum": ["RETRY", "CANCEL"]}, "reason": {"type": "string", "minLength": 3, "maxLength": 500}}}}}}
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/production/cards:
    get:
      tags: [Platform]
      operationId: platformCards
      summary: "Paginated lifecycle registry; optional school_id, status and student-name search"
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/production/cards/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    patch:
      tags: [Platform]
      operationId: platformCardStatus
      summary: "Change status with production and replacement checks"
      security: [{ bearerAuth: [] }]
      requestBody: {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["status", "reason"], "properties": {"status": {"type": "string", "enum": ["ACTIVE", "LOST", "BLOCKED", "EXPIRED"]}, "reason": {"type": "string", "minLength": 3, "maxLength": 500}}}}}}
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    schoolId:
      type: apiKey
      in: header
      name: X-School-Id
      description: UUID of the active school membership.
    deviceAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Per-device credential formatted as `Device <token>`.
    deviceId:
      type: apiKey
      in: header
      name: X-Device-Id
      description: Registered device UUID; school context is derived server-side.
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Search:
      name: search
      in: query
      schema: { type: string, minLength: 1, maxLength: 100, pattern: '^[\p{L}\p{N}\s.-]+$' }
  responses:
...
    AuthRequired:
      description: Missing, expired, or invalid bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Forbidden:
      description: Tenant membership or permission is missing
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found in the active tenant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: Resource violates a uniqueness constraint
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ValidationError:
      description: Request payload, parameters, or query are invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: {}
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [success, error, request_id]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [AUTH_REQUIRED, AUTH_INVALID, INVALID_CREDENTIALS, SESSION_REFRESH_FAILED, FORBIDDEN, TENANT_REQUIRED, SCHOOL_NOT_FOUND, ACADEMIC_YEAR_NOT_FOUND, CLASS_NOT_FOUND, STUDENT_NOT_FOUND, CARD_NOT_FOUND, DEVICE_NOT_REGISTERED, DEVICE_AUTH_INVALID, DEVICE_NOT_GATE, ATTENDANCE_DISABLED, CARD_BLOCKED, STUDENT_INACTIVE, GATE_RULE_VIOLATION, PARENT_LINK_INVALID, CHILD_NOT_LINKED, RESOURCE_NOT_FOUND, INVALID_JSON, PAYLOAD_TOO_LARGE, VALIDATION_ERROR, CONFLICT, RATE_LIMITED, ROUTE_NOT_FOUND, DATABASE_ERROR, SERVER_UNAVAILABLE, INTERNAL_ERROR]
            message: { type: string }
            details: {}
        request_id: { type: string }
    PaginationMeta:
      type: object
      required: [page, page_size, total]
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
...
    Class:
      type: object
      required: [id, school_id, academic_year_id, code, name, is_active]
      properties:
        id: { type: string, format: uuid }
        school_id: { type: string, format: uuid }
        academic_year_id: { type: string, format: uuid }
        code: { type: string, maxLength: 50 }
        name: { type: string, maxLength: 120 }
        grade_level: { type: [integer, 'null'], minimum: 1, maximum: 12 }
        homeroom_teacher_user_id: { type: [string, 'null'], format: uuid }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ClassPayload:
      type: object
      additionalProperties: false
      properties:
        academic_year_id: { type: string, format: uuid }
        code: { type: string, minLength: 1, maxLength: 50 }
        name: { type: string, minLength: 1, maxLength: 120 }
        grade_level: { type: [integer, 'null'], minimum: 1, maximum: 12 }
        homeroom_teacher_user_id: { type: [string, 'null'], format: uuid }
        is_active: { type: boolean }
    ClassCreate:
      allOf:
        - $ref: '#/components/schemas/ClassPayload'
      required: [academic_year_id, code, name]
...
    ClassResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Class' }
...
    CardStatus: { type: string, enum: [ACTIVE, LOST, BLOCKED, REPLACED, EXPIRED] }
    Card:
      type: object
      required: [id, school_id, student_id, card_uid, card_serial, qr_key, status]
      properties:
        id: { type: string, format: uuid }
        school_id: { type: string, format: uuid }
        student_id: { type: string, format: uuid }
        card_uid: { type: [string, "null"] }
        production_status: { type: string, enum: [LEGACY, DRAFT, PRINTED, READY_TO_WRITE, WRITING, VERIFIED, FAILED, CANCELLED] }
        card_serial: { type: string }
        qr_key: { type: string, description: Opaque random identifier containing no PII. }
        status: { $ref: '#/components/schemas/CardStatus' }
        issued_at: { type: [string, 'null'], format: date-time }
        revoked_at: { type: [string, 'null'], format: date-time }
        expires_at: { type: [string, 'null'], format: date-time }
...
    CardUpdate:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        reason: { type: string, minLength: 3, maxLength: 500 }
        status: { $ref: '#/components/schemas/CardStatus' }
        expires_at: { type: [string, 'null'], format: date-time }
    CardResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - { type: object, properties: { data: { $ref: '#/components/schemas/Card' } } }
    CardListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/Card' } }
            meta: { $ref: '#/components/schemas/PaginationMeta' }
...